# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CacheLab demo server API
  version: 0.1.0
  description: |
    REST API and metrics stream of the CacheLab demo server (SPEC 8.2).
    Errors are RFC 7807 `application/problem+json` documents.

    **Stub (Step 1).** Every endpoint is listed; `x-cachelab-step` says which build step
    delivers it. Implemented today: `GET /api/health`, `GET /api/metrics/stream`.
    Live, generated docs: `/swagger-ui` on a running server.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/simulations:
    post:
      tags: [simulations]
      summary: Start a workload on a group (stops any running simulation first; one runs at a time)
      description: |
        The same key sequence is applied to every cache of the group, in the same order.
        Progress is reported in the `simulation` field of the metrics stream.
      x-cachelab-step: 3
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SimulationRequest' }
      responses:
        '200':
          description: Started
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: string, example: s-12 }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /api/simulations/{id}:
    delete:
      tags: [simulations]
      summary: Stop a simulation (no-op if it already finished)
      x-cachelab-step: 3
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Stopped }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
...
  responses:
    Problem:
      description: RFC 7807 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
...
    Pattern:
      type: string
      enum: [UNIFORM, ZIPF, SCAN_POLLUTION, LOOP, SHIFTING_HOTSPOT, TTL_BURST, FORMULARY, PROVIDER_DIRECTORY]
...
    SimulationRequest:
      type: object
      required: [group, pattern]
      properties:
        group: { type: string, description: an existing group, e.g. demo }
        pattern: { $ref: '#/components/schemas/Pattern' }
        opsPerSec: { type: integer, minimum: 0, maximum: 200000, default: 5000, description: 0 = unthrottled }
        readRatio: { type: number, minimum: 0, maximum: 1, default: 0.9 }
        durationSec: { type: integer, minimum: 1, maximum: 3600, default: 60 }
        seed: { type: integer, default: 42 }
        params:
          type: object
          additionalProperties: true
          description: Pattern overrides, e.g. keySpace, zipfS (see SPEC 9.1 defaults)
...
    ProblemDetail:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }